scope persistence refactor outside orchestration
