--- apps/server/src/persistence/OrchestrationCommandReceipts.ts
@@ -1,150 +1,0 @@
-import * as Context from "effect/Context";
-import * as Effect from "effect/Effect";
-import * as Layer from "effect/Layer";
-import * as Option from "effect/Option";
-import * as Schema from "effect/Schema";
-import * as SqlClient from "effect/unstable/sql/SqlClient";
-import * as SqlSchema from "effect/unstable/sql/SqlSchema";
-
-import {
-  CommandId,
-  IsoDateTime,
-  NonNegativeInt,
-  OrchestrationAggregateKind,
-  OrchestrationCommandReceiptStatus,
-  ProjectId,
-  ThreadId,
-} from "@t3tools/contracts";
-
-import {
-  type OrchestrationCommandReceiptRepositoryError,
-  toPersistenceSqlError,
-} from "./Errors.ts";
-
-/**
- * OrchestrationCommandReceiptRepository - Repository interface for command receipts.
- *
- * Owns persistence operations for deduplication and status tracking of
- * orchestration command handling.
- *
- * @module OrchestrationCommandReceiptRepository
- */
-
-export const OrchestrationCommandReceipt = Schema.Struct({
-  commandId: CommandId,
-  aggregateKind: OrchestrationAggregateKind,
-  aggregateId: Schema.Union([ProjectId, ThreadId]),
-  acceptedAt: IsoDateTime,
-  resultSequence: NonNegativeInt,
-  status: OrchestrationCommandReceiptStatus,
-  error: Schema.NullOr(Schema.String),
-});
-export type OrchestrationCommandReceipt = typeof OrchestrationCommandReceipt.Type;
-
-export const GetByCommandIdInput = Schema.Struct({
-  commandId: CommandId,
-});
-export type GetByCommandIdInput = typeof GetByCommandIdInput.Type;
-
-/**
- * OrchestrationCommandReceiptRepository - Service tag for command receipt persistence.
- */
-export class OrchestrationCommandReceiptRepository extends Context.Service<
-  OrchestrationCommandReceiptRepository,
-  {
-    /**
-     * Insert or replace a command receipt row.
-     *
-     * Upserts by `commandId` for idempotent command-result tracking.
-     */
-    readonly upsert: (
-      receipt: OrchestrationCommandReceipt,
-    ) => Effect.Effect<void, OrchestrationCommandReceiptRepositoryError>;
-
-    /**
-     * Read a command receipt by command id.
-     */
-    readonly getByCommandId: (
-      input: GetByCommandIdInput,
-    ) => Effect.Effect<
-      Option.Option<OrchestrationCommandReceipt>,
-      OrchestrationCommandReceiptRepositoryError
-    >;
-  }
->()("t3/persistence/OrchestrationCommandReceipts/OrchestrationCommandReceiptRepository") {}
-
-export const make = Effect.gen(function* () {
-  const sql = yield* SqlClient.SqlClient;
-
-  const upsertReceiptRow = SqlSchema.void({
-    Request: OrchestrationCommandReceipt,
-    execute: (receipt) =>
-      sql`
-        INSERT INTO orchestration_command_receipts (
-          command_id,
-          aggregate_kind,
-          aggregate_id,
-          accepted_at,
-          result_sequence,
-          status,
-          error
-        )
-        VALUES (
-          ${receipt.commandId},
-          ${receipt.aggregateKind},
-          ${receipt.aggregateId},
-          ${receipt.acceptedAt},
-          ${receipt.resultSequence},
-          ${receipt.status},
-          ${receipt.error}
-        )
-        ON CONFLICT (command_id)
-        DO UPDATE SET
-          aggregate_kind = excluded.aggregate_kind,
-          aggregate_id = excluded.aggregate_id,
-          accepted_at = excluded.accepted_at,
-          result_sequence = excluded.result_sequence,
-          status = excluded.status,
-          error = excluded.error
-      `,
-  });
-
-  const findReceiptByCommandId = SqlSchema.findOneOption({
-    Request: GetByCommandIdInput,
-    Result: OrchestrationCommandReceipt,
-    execute: ({ commandId }) =>
-      sql`
-        SELECT
-          command_id AS "commandId",
-          aggregate_kind AS "aggregateKind",
-          aggregate_id AS "aggregateId",
-          accepted_at AS "acceptedAt",
-          result_sequence AS "resultSequence",
-          status,
-          error
-        FROM orchestration_command_receipts
-        WHERE command_id = ${commandId}
-      `,
-  });
-
-  const upsert: OrchestrationCommandReceiptRepository["Service"]["upsert"] = (receipt) =>
-    upsertReceiptRow(receipt).pipe(
-      Effect.mapError(toPersistenceSqlError("OrchestrationCommandReceiptRepository.upsert:query")),
-    );
-
-  const getByCommandId: OrchestrationCommandReceiptRepository["Service"]["getByCommandId"] = (
-    input,
-  ) =>
-    findReceiptByCommandId(input).pipe(
-      Effect.mapError(
-        toPersistenceSqlError("OrchestrationCommandReceiptRepository.getByCommandId:query"),
-      ),
-    );
-
-  return {
-    upsert,
-    getByCommandId,
-  } satisfies OrchestrationCommandReceiptRepository["Service"];
-});
-
-export const layer = Layer.effect(OrchestrationCommandReceiptRepository, make);
