--- apps/server/src/persistence/ProjectionThreadActivities.ts
@@ -1,230 +1,0 @@
-import * as Context from "effect/Context";
-import * as Effect from "effect/Effect";
-import * as Layer from "effect/Layer";
-import * as Schema from "effect/Schema";
-import * as Struct from "effect/Struct";
-import * as SqlClient from "effect/unstable/sql/SqlClient";
-import * as SqlSchema from "effect/unstable/sql/SqlSchema";
-
-import {
-  EventId,
-  IsoDateTime,
-  NonNegativeInt,
-  OrchestrationThreadActivityTone,
-  ThreadId,
-  TurnId,
-} from "@t3tools/contracts";
-
-import {
-  type ProjectionRepositoryError,
-  toPersistenceDecodeError,
-  toPersistenceSqlError,
-} from "./Errors.ts";
-
-/**
- * ProjectionThreadActivityRepository - Projection repository interface for thread activity.
- *
- * Owns persistence operations for activity timeline entries projected from
- * orchestration events.
- *
- * @module ProjectionThreadActivityRepository
- */
-
-export const ProjectionThreadActivity = Schema.Struct({
-  activityId: EventId,
-  threadId: ThreadId,
-  turnId: Schema.NullOr(TurnId),
-  tone: OrchestrationThreadActivityTone,
-  kind: Schema.String,
-  summary: Schema.String,
-  payload: Schema.Unknown,
-  sequence: Schema.optional(NonNegativeInt),
-  createdAt: IsoDateTime,
-});
-export type ProjectionThreadActivity = typeof ProjectionThreadActivity.Type;
-
-export const ListProjectionThreadActivitiesInput = Schema.Struct({
-  threadId: ThreadId,
-});
-export type ListProjectionThreadActivitiesInput = typeof ListProjectionThreadActivitiesInput.Type;
-
-export const DeleteProjectionThreadActivitiesInput = Schema.Struct({
-  threadId: ThreadId,
-});
-export type DeleteProjectionThreadActivitiesInput =
-  typeof DeleteProjectionThreadActivitiesInput.Type;
-
-/**
- * ProjectionThreadActivityRepository - Service tag for thread activity persistence.
- */
-export class ProjectionThreadActivityRepository extends Context.Service<
-  ProjectionThreadActivityRepository,
-  {
-    /**
-     * Insert or replace a projected thread activity row.
-     *
-     * Upserts by `activityId` and JSON-encodes payload.
-     */
-    readonly upsert: (
-      row: ProjectionThreadActivity,
-    ) => Effect.Effect<void, ProjectionRepositoryError>;
-
-    /**
-     * List projected thread activity rows for a thread.
-     *
-     * Returned in ascending runtime sequence order (or creation order when
-     * sequence is unavailable).
-     */
-    readonly listByThreadId: (
-      input: ListProjectionThreadActivitiesInput,
-    ) => Effect.Effect<ReadonlyArray<ProjectionThreadActivity>, ProjectionRepositoryError>;
-
-    /**
-     * Delete projected thread activity rows by thread.
-     */
-    readonly deleteByThreadId: (
-      input: DeleteProjectionThreadActivitiesInput,
-    ) => Effect.Effect<void, ProjectionRepositoryError>;
-  }
->()("t3/persistence/ProjectionThreadActivities/ProjectionThreadActivityRepository") {}
-
-const ProjectionThreadActivityDbRowSchema = ProjectionThreadActivity.mapFields(
-  Struct.assign({
-    payload: Schema.fromJsonString(Schema.Unknown),
-    sequence: Schema.NullOr(NonNegativeInt),
-  }),
-);
-
-function toPersistenceSqlOrDecodeError(sqlOperation: string, decodeOperation: string) {
-  return (cause: unknown) =>
-    Schema.isSchemaError(cause)
-      ? toPersistenceDecodeError(decodeOperation)(cause)
-      : toPersistenceSqlError(sqlOperation)(cause);
-}
-
-export const make = Effect.gen(function* () {
-  const sql = yield* SqlClient.SqlClient;
-
-  const upsertProjectionThreadActivityRow = SqlSchema.void({
-    Request: ProjectionThreadActivity,
-    execute: (row) =>
-      sql`
-            INSERT INTO projection_thread_activities (
-              activity_id,
-              thread_id,
-              turn_id,
-              tone,
-              kind,
-              summary,
-              payload_json,
-              sequence,
-              created_at
-            )
-            VALUES (
-              ${row.activityId},
-              ${row.threadId},
-              ${row.turnId},
-              ${row.tone},
-              ${row.kind},
-              ${row.summary},
-              ${JSON.stringify(row.payload)},
-              ${row.sequence ?? null},
-              ${row.createdAt}
-            )
-            ON CONFLICT (activity_id)
-            DO UPDATE SET
-              thread_id = excluded.thread_id,
-              turn_id = excluded.turn_id,
-              tone = excluded.tone,
-              kind = excluded.kind,
-              summary = excluded.summary,
-              payload_json = excluded.payload_json,
-              sequence = excluded.sequence,
-              created_at = excluded.created_at
-          `,
-  });
-
-  const listProjectionThreadActivityRows = SqlSchema.findAll({
-    Request: ListProjectionThreadActivitiesInput,
-    Result: ProjectionThreadActivityDbRowSchema,
-    execute: ({ threadId }) =>
-      sql`
-        SELECT
-          activity_id AS "activityId",
-          thread_id AS "threadId",
-          turn_id AS "turnId",
-          tone,
-          kind,
-          summary,
-          payload_json AS "payload",
-          sequence,
-          created_at AS "createdAt"
-        FROM projection_thread_activities
-        WHERE thread_id = ${threadId}
-        ORDER BY
-          CASE WHEN sequence IS NULL THEN 0 ELSE 1 END ASC,
-          sequence ASC,
-          created_at ASC,
-          activity_id ASC
-      `,
-  });
-
-  const deleteProjectionThreadActivityRows = SqlSchema.void({
-    Request: DeleteProjectionThreadActivitiesInput,
-    execute: ({ threadId }) =>
-      sql`
-        DELETE FROM projection_thread_activities
-        WHERE thread_id = ${threadId}
-      `,
-  });
-
-  const upsert: ProjectionThreadActivityRepository["Service"]["upsert"] = (row) =>
-    upsertProjectionThreadActivityRow(row).pipe(
-      Effect.mapError(
-        toPersistenceSqlOrDecodeError(
-          "ProjectionThreadActivityRepository.upsert:query",
-          "ProjectionThreadActivityRepository.upsert:encodeRequest",
-        ),
-      ),
-    );
-
-  const listByThreadId: ProjectionThreadActivityRepository["Service"]["listByThreadId"] = (input) =>
-    listProjectionThreadActivityRows(input).pipe(
-      Effect.mapError(
-        toPersistenceSqlOrDecodeError(
-          "ProjectionThreadActivityRepository.listByThreadId:query",
-          "ProjectionThreadActivityRepository.listByThreadId:decodeRows",
-        ),
-      ),
-      Effect.map((rows) =>
-        rows.map((row) => ({
-          activityId: row.activityId,
-          threadId: row.threadId,
-          turnId: row.turnId,
-          tone: row.tone,
-          kind: row.kind,
-          summary: row.summary,
-          payload: row.payload,
-          ...(row.sequence !== null ? { sequence: row.sequence } : {}),
-          createdAt: row.createdAt,
-        })),
-      ),
-    );
-
-  const deleteByThreadId: ProjectionThreadActivityRepository["Service"]["deleteByThreadId"] = (
-    input,
-  ) =>
-    deleteProjectionThreadActivityRows(input).pipe(
-      Effect.mapError(
-        toPersistenceSqlError("ProjectionThreadActivityRepository.deleteByThreadId:query"),
-      ),
-    );
-
-  return {
-    upsert,
-    listByThreadId,
-    deleteByThreadId,
-  } satisfies ProjectionThreadActivityRepository["Service"];
-});
-
-export const layer = Layer.effect(ProjectionThreadActivityRepository, make);
