--- apps/server/src/persistence/ProjectionThreadMessages.ts
@@ -1,265 +1,0 @@
-import * as Context from "effect/Context";
-import * as Effect from "effect/Effect";
-import * as Layer from "effect/Layer";
-import * as Option from "effect/Option";
-import * as Schema from "effect/Schema";
-import * as Struct from "effect/Struct";
-import * as SqlClient from "effect/unstable/sql/SqlClient";
-import * as SqlSchema from "effect/unstable/sql/SqlSchema";
-
-import {
-  ChatAttachment,
-  IsoDateTime,
-  MessageId,
-  OrchestrationMessageRole,
-  ThreadId,
-  TurnId,
-} from "@t3tools/contracts";
-
-import { type ProjectionRepositoryError, toPersistenceSqlError } from "./Errors.ts";
-
-/**
- * ProjectionThreadMessageRepository - Projection repository interface for messages.
- *
- * Owns persistence operations for projected thread messages rendered in the
- * orchestration read model.
- *
- * @module ProjectionThreadMessageRepository
- */
-
-export const ProjectionThreadMessage = Schema.Struct({
-  messageId: MessageId,
-  threadId: ThreadId,
-  turnId: Schema.NullOr(TurnId),
-  role: OrchestrationMessageRole,
-  text: Schema.String,
-  attachments: Schema.optional(Schema.Array(ChatAttachment)),
-  isStreaming: Schema.Boolean,
-  createdAt: IsoDateTime,
-  updatedAt: IsoDateTime,
-});
-export type ProjectionThreadMessage = typeof ProjectionThreadMessage.Type;
-
-export const ListProjectionThreadMessagesInput = Schema.Struct({
-  threadId: ThreadId,
-});
-export type ListProjectionThreadMessagesInput = typeof ListProjectionThreadMessagesInput.Type;
-
-export const GetProjectionThreadMessageInput = Schema.Struct({
-  messageId: MessageId,
-});
-export type GetProjectionThreadMessageInput = typeof GetProjectionThreadMessageInput.Type;
-
-export const DeleteProjectionThreadMessagesInput = Schema.Struct({
-  threadId: ThreadId,
-});
-export type DeleteProjectionThreadMessagesInput = typeof DeleteProjectionThreadMessagesInput.Type;
-
-/**
- * ProjectionThreadMessageRepository - Service tag for message projection persistence.
- */
-export class ProjectionThreadMessageRepository extends Context.Service<
-  ProjectionThreadMessageRepository,
-  {
-    /**
-     * Insert or replace a projected thread message row.
-     *
-     * Upserts by `messageId`.
-     */
-    readonly upsert: (
-      message: ProjectionThreadMessage,
-    ) => Effect.Effect<void, ProjectionRepositoryError>;
-
-    /**
-     * Read a projected thread message by id.
-     */
-    readonly getByMessageId: (
-      input: GetProjectionThreadMessageInput,
-    ) => Effect.Effect<Option.Option<ProjectionThreadMessage>, ProjectionRepositoryError>;
-
-    /**
-     * List projected thread messages for a thread.
-     *
-     * Returned in ascending creation order.
-     */
-    readonly listByThreadId: (
-      input: ListProjectionThreadMessagesInput,
-    ) => Effect.Effect<ReadonlyArray<ProjectionThreadMessage>, ProjectionRepositoryError>;
-
-    /**
-     * Delete projected thread messages by thread.
-     */
-    readonly deleteByThreadId: (
-      input: DeleteProjectionThreadMessagesInput,
-    ) => Effect.Effect<void, ProjectionRepositoryError>;
-  }
->()("t3/persistence/ProjectionThreadMessages/ProjectionThreadMessageRepository") {}
-
-const ProjectionThreadMessageDbRowSchema = ProjectionThreadMessage.mapFields(
-  Struct.assign({
-    isStreaming: Schema.Number,
-    attachments: Schema.NullOr(Schema.fromJsonString(Schema.Array(ChatAttachment))),
-  }),
-);
-
-function toProjectionThreadMessage(
-  row: Schema.Schema.Type<typeof ProjectionThreadMessageDbRowSchema>,
-): ProjectionThreadMessage {
-  return {
-    messageId: row.messageId,
-    threadId: row.threadId,
-    turnId: row.turnId,
-    role: row.role,
-    text: row.text,
-    isStreaming: row.isStreaming === 1,
-    createdAt: row.createdAt,
-    updatedAt: row.updatedAt,
-    ...(row.attachments !== null ? { attachments: row.attachments } : {}),
-  };
-}
-
-export const make = Effect.gen(function* () {
-  const sql = yield* SqlClient.SqlClient;
-
-  const upsertProjectionThreadMessageRow = SqlSchema.void({
-    Request: ProjectionThreadMessage,
-    execute: (row) => {
-      const nextAttachmentsJson =
-        row.attachments !== undefined ? JSON.stringify(row.attachments) : null;
-      return sql`
-        INSERT INTO projection_thread_messages (
-          message_id,
-          thread_id,
-          turn_id,
-          role,
-          text,
-          attachments_json,
-          is_streaming,
-          created_at,
-          updated_at
-        )
-        VALUES (
-          ${row.messageId},
-          ${row.threadId},
-          ${row.turnId},
-          ${row.role},
-          ${row.text},
-          COALESCE(
-            ${nextAttachmentsJson},
-            (
-              SELECT attachments_json
-              FROM projection_thread_messages
-              WHERE message_id = ${row.messageId}
-            )
-          ),
-          ${row.isStreaming ? 1 : 0},
-          ${row.createdAt},
-          ${row.updatedAt}
-        )
-        ON CONFLICT (message_id)
-        DO UPDATE SET
-          thread_id = excluded.thread_id,
-          turn_id = excluded.turn_id,
-          role = excluded.role,
-          text = excluded.text,
-          attachments_json = COALESCE(
-            excluded.attachments_json,
-            projection_thread_messages.attachments_json
-          ),
-          is_streaming = excluded.is_streaming,
-          created_at = excluded.created_at,
-          updated_at = excluded.updated_at
-      `;
-    },
-  });
-
-  const getProjectionThreadMessageRow = SqlSchema.findOneOption({
-    Request: GetProjectionThreadMessageInput,
-    Result: ProjectionThreadMessageDbRowSchema,
-    execute: ({ messageId }) =>
-      sql`
-        SELECT
-          message_id AS "messageId",
-          thread_id AS "threadId",
-          turn_id AS "turnId",
-          role,
-          text,
-          attachments_json AS "attachments",
-          is_streaming AS "isStreaming",
-          created_at AS "createdAt",
-          updated_at AS "updatedAt"
-        FROM projection_thread_messages
-        WHERE message_id = ${messageId}
-        LIMIT 1
-      `,
-  });
-
-  const listProjectionThreadMessageRows = SqlSchema.findAll({
-    Request: ListProjectionThreadMessagesInput,
-    Result: ProjectionThreadMessageDbRowSchema,
-    execute: ({ threadId }) =>
-      sql`
-        SELECT
-          message_id AS "messageId",
-          thread_id AS "threadId",
-          turn_id AS "turnId",
-          role,
-          text,
-          attachments_json AS "attachments",
-          is_streaming AS "isStreaming",
-          created_at AS "createdAt",
-          updated_at AS "updatedAt"
-        FROM projection_thread_messages
-        WHERE thread_id = ${threadId}
-        ORDER BY created_at ASC, message_id ASC
-      `,
-  });
-
-  const deleteProjectionThreadMessageRows = SqlSchema.void({
-    Request: DeleteProjectionThreadMessagesInput,
-    execute: ({ threadId }) =>
-      sql`
-        DELETE FROM projection_thread_messages
-        WHERE thread_id = ${threadId}
-      `,
-  });
-
-  const upsert: ProjectionThreadMessageRepository["Service"]["upsert"] = (row) =>
-    upsertProjectionThreadMessageRow(row).pipe(
-      Effect.mapError(toPersistenceSqlError("ProjectionThreadMessageRepository.upsert:query")),
-    );
-
-  const getByMessageId: ProjectionThreadMessageRepository["Service"]["getByMessageId"] = (input) =>
-    getProjectionThreadMessageRow(input).pipe(
-      Effect.mapError(
-        toPersistenceSqlError("ProjectionThreadMessageRepository.getByMessageId:query"),
-      ),
-      Effect.map(Option.map(toProjectionThreadMessage)),
-    );
-
-  const listByThreadId: ProjectionThreadMessageRepository["Service"]["listByThreadId"] = (input) =>
-    listProjectionThreadMessageRows(input).pipe(
-      Effect.mapError(
-        toPersistenceSqlError("ProjectionThreadMessageRepository.listByThreadId:query"),
-      ),
-      Effect.map((rows) => rows.map(toProjectionThreadMessage)),
-    );
-
-  const deleteByThreadId: ProjectionThreadMessageRepository["Service"]["deleteByThreadId"] = (
-    input,
-  ) =>
-    deleteProjectionThreadMessageRows(input).pipe(
-      Effect.mapError(
-        toPersistenceSqlError("ProjectionThreadMessageRepository.deleteByThreadId:query"),
-      ),
-    );
-
-  return {
-    upsert,
-    getByMessageId,
-    listByThreadId,
-    deleteByThreadId,
-  } satisfies ProjectionThreadMessageRepository["Service"];
-});
-
-export const layer = Layer.effect(ProjectionThreadMessageRepository, make);
