--- apps/server/src/persistence/ProjectionThreadSessions.ts
@@ -1,177 +1,0 @@
-import * as Context from "effect/Context";
-import * as Effect from "effect/Effect";
-import * as Layer from "effect/Layer";
-import * as Option from "effect/Option";
-import * as Schema from "effect/Schema";
-import * as SqlClient from "effect/unstable/sql/SqlClient";
-import * as SqlSchema from "effect/unstable/sql/SqlSchema";
-
-import {
-  IsoDateTime,
-  OrchestrationSessionStatus,
-  ProviderInstanceId,
-  RuntimeMode,
-  ThreadId,
-  TurnId,
-} from "@t3tools/contracts";
-
-import { type ProjectionRepositoryError, toPersistenceSqlError } from "./Errors.ts";
-
-/**
- * ProjectionThreadSessionRepository - Repository interface for thread sessions.
- *
- * Owns persistence operations for projected provider-session linkage and
- * runtime status for each thread.
- *
- * @module ProjectionThreadSessionRepository
- */
-
-export const ProjectionThreadSession = Schema.Struct({
-  threadId: ThreadId,
-  status: OrchestrationSessionStatus,
-  providerName: Schema.NullOr(Schema.String),
-  providerInstanceId: Schema.NullOr(ProviderInstanceId),
-  runtimeMode: RuntimeMode,
-  activeTurnId: Schema.NullOr(TurnId),
-  lastError: Schema.NullOr(Schema.String),
-  updatedAt: IsoDateTime,
-});
-export type ProjectionThreadSession = typeof ProjectionThreadSession.Type;
-
-export const GetProjectionThreadSessionInput = Schema.Struct({
-  threadId: ThreadId,
-});
-export type GetProjectionThreadSessionInput = typeof GetProjectionThreadSessionInput.Type;
-
-export const DeleteProjectionThreadSessionInput = Schema.Struct({
-  threadId: ThreadId,
-});
-export type DeleteProjectionThreadSessionInput = typeof DeleteProjectionThreadSessionInput.Type;
-
-/**
- * ProjectionThreadSessionRepository - Service tag for thread-session persistence.
- */
-export class ProjectionThreadSessionRepository extends Context.Service<
-  ProjectionThreadSessionRepository,
-  {
-    /**
-     * Insert or replace a projected thread-session row.
-     *
-     * Upserts by `threadId`.
-     */
-    readonly upsert: (
-      row: ProjectionThreadSession,
-    ) => Effect.Effect<void, ProjectionRepositoryError>;
-
-    /**
-     * Read projected thread-session state by thread id.
-     */
-    readonly getByThreadId: (
-      input: GetProjectionThreadSessionInput,
-    ) => Effect.Effect<Option.Option<ProjectionThreadSession>, ProjectionRepositoryError>;
-
-    /**
-     * Delete projected thread-session state by thread id.
-     */
-    readonly deleteByThreadId: (
-      input: DeleteProjectionThreadSessionInput,
-    ) => Effect.Effect<void, ProjectionRepositoryError>;
-  }
->()("t3/persistence/ProjectionThreadSessions/ProjectionThreadSessionRepository") {}
-
-export const make = Effect.gen(function* () {
-  const sql = yield* SqlClient.SqlClient;
-
-  const upsertProjectionThreadSessionRow = SqlSchema.void({
-    Request: ProjectionThreadSession,
-    execute: (row) =>
-      sql`
-        INSERT INTO projection_thread_sessions (
-          thread_id,
-          status,
-          provider_name,
-          provider_instance_id,
-          runtime_mode,
-          active_turn_id,
-          last_error,
-          updated_at
-        )
-        VALUES (
-          ${row.threadId},
-          ${row.status},
-          ${row.providerName},
-          ${row.providerInstanceId},
-          ${row.runtimeMode},
-          ${row.activeTurnId},
-          ${row.lastError},
-          ${row.updatedAt}
-        )
-        ON CONFLICT (thread_id)
-        DO UPDATE SET
-          status = excluded.status,
-          provider_name = excluded.provider_name,
-          provider_instance_id = excluded.provider_instance_id,
-          runtime_mode = excluded.runtime_mode,
-          active_turn_id = excluded.active_turn_id,
-          last_error = excluded.last_error,
-          updated_at = excluded.updated_at
-      `,
-  });
-
-  const getProjectionThreadSessionRow = SqlSchema.findOneOption({
-    Request: GetProjectionThreadSessionInput,
-    Result: ProjectionThreadSession,
-    execute: ({ threadId }) =>
-      sql`
-        SELECT
-          thread_id AS "threadId",
-          status,
-          provider_name AS "providerName",
-          provider_instance_id AS "providerInstanceId",
-          runtime_mode AS "runtimeMode",
-          active_turn_id AS "activeTurnId",
-          last_error AS "lastError",
-          updated_at AS "updatedAt"
-        FROM projection_thread_sessions
-        WHERE thread_id = ${threadId}
-      `,
-  });
-
-  const deleteProjectionThreadSessionRow = SqlSchema.void({
-    Request: DeleteProjectionThreadSessionInput,
-    execute: ({ threadId }) =>
-      sql`
-        DELETE FROM projection_thread_sessions
-        WHERE thread_id = ${threadId}
-      `,
-  });
-
-  const upsert: ProjectionThreadSessionRepository["Service"]["upsert"] = (row) =>
-    upsertProjectionThreadSessionRow(row).pipe(
-      Effect.mapError(toPersistenceSqlError("ProjectionThreadSessionRepository.upsert:query")),
-    );
-
-  const getByThreadId: ProjectionThreadSessionRepository["Service"]["getByThreadId"] = (input) =>
-    getProjectionThreadSessionRow(input).pipe(
-      Effect.mapError(
-        toPersistenceSqlError("ProjectionThreadSessionRepository.getByThreadId:query"),
-      ),
-    );
-
-  const deleteByThreadId: ProjectionThreadSessionRepository["Service"]["deleteByThreadId"] = (
-    input,
-  ) =>
-    deleteProjectionThreadSessionRow(input).pipe(
-      Effect.mapError(
-        toPersistenceSqlError("ProjectionThreadSessionRepository.deleteByThreadId:query"),
-      ),
-    );
-
-  return {
-    upsert,
-    getByThreadId,
-    deleteByThreadId,
-  } satisfies ProjectionThreadSessionRepository["Service"];
-});
-
-export const layer = Layer.effect(ProjectionThreadSessionRepository, make);
