--- apps/server/src/persistence/ProjectionThreads.ts
@@ -1,273 +1,0 @@
-import * as Context from "effect/Context";
-import * as Effect from "effect/Effect";
-import * as Layer from "effect/Layer";
-import * as Option from "effect/Option";
-import * as Schema from "effect/Schema";
-import * as Struct from "effect/Struct";
-import * as SqlClient from "effect/unstable/sql/SqlClient";
-import * as SqlSchema from "effect/unstable/sql/SqlSchema";
-
-import {
-  IsoDateTime,
-  ModelSelection,
-  NonNegativeInt,
-  ProjectId,
-  ProviderInteractionMode,
-  RuntimeMode,
-  ThreadId,
-  TurnId,
-} from "@t3tools/contracts";
-
-import { type ProjectionRepositoryError, toPersistenceSqlError } from "./Errors.ts";
-
-/**
- * ProjectionThreadRepository - Projection repository interface for threads.
- *
- * Owns persistence operations for projected thread records in the
- * orchestration read model.
- *
- * @module ProjectionThreadRepository
- */
-
-export const ProjectionThread = Schema.Struct({
-  threadId: ThreadId,
-  projectId: ProjectId,
-  title: Schema.String,
-  modelSelection: ModelSelection,
-  runtimeMode: RuntimeMode,
-  interactionMode: ProviderInteractionMode,
-  branch: Schema.NullOr(Schema.String),
-  worktreePath: Schema.NullOr(Schema.String),
-  latestTurnId: Schema.NullOr(TurnId),
-  createdAt: IsoDateTime,
-  updatedAt: IsoDateTime,
-  archivedAt: Schema.NullOr(IsoDateTime),
-  latestUserMessageAt: Schema.NullOr(IsoDateTime),
-  pendingApprovalCount: NonNegativeInt,
-  pendingUserInputCount: NonNegativeInt,
-  hasActionableProposedPlan: NonNegativeInt,
-  deletedAt: Schema.NullOr(IsoDateTime),
-});
-export type ProjectionThread = typeof ProjectionThread.Type;
-
-export const GetProjectionThreadInput = Schema.Struct({
-  threadId: ThreadId,
-});
-export type GetProjectionThreadInput = typeof GetProjectionThreadInput.Type;
-
-export const DeleteProjectionThreadInput = Schema.Struct({
-  threadId: ThreadId,
-});
-export type DeleteProjectionThreadInput = typeof DeleteProjectionThreadInput.Type;
-
-export const ListProjectionThreadsByProjectInput = Schema.Struct({
-  projectId: ProjectId,
-});
-export type ListProjectionThreadsByProjectInput = typeof ListProjectionThreadsByProjectInput.Type;
-
-/**
- * ProjectionThreadRepository - Service tag for thread projection persistence.
- */
-export class ProjectionThreadRepository extends Context.Service<
-  ProjectionThreadRepository,
-  {
-    /**
-     * Insert or replace a projected thread row.
-     *
-     * Upserts by `threadId`.
-     */
-    readonly upsert: (thread: ProjectionThread) => Effect.Effect<void, ProjectionRepositoryError>;
-
-    /**
-     * Read a projected thread row by id.
-     */
-    readonly getById: (
-      input: GetProjectionThreadInput,
-    ) => Effect.Effect<Option.Option<ProjectionThread>, ProjectionRepositoryError>;
-
-    /**
-     * List projected threads for a project.
-     *
-     * Returned in deterministic creation order.
-     */
-    readonly listByProjectId: (
-      input: ListProjectionThreadsByProjectInput,
-    ) => Effect.Effect<ReadonlyArray<ProjectionThread>, ProjectionRepositoryError>;
-
-    /**
-     * Soft-delete a projected thread row by id.
-     */
-    readonly deleteById: (
-      input: DeleteProjectionThreadInput,
-    ) => Effect.Effect<void, ProjectionRepositoryError>;
-  }
->()("t3/persistence/ProjectionThreads/ProjectionThreadRepository") {}
-
-const ProjectionThreadDbRow = ProjectionThread.mapFields(
-  Struct.assign({
-    modelSelection: Schema.fromJsonString(ModelSelection),
-  }),
-);
-type ProjectionThreadDbRow = typeof ProjectionThreadDbRow.Type;
-
-export const make = Effect.gen(function* () {
-  const sql = yield* SqlClient.SqlClient;
-
-  const upsertProjectionThreadRow = SqlSchema.void({
-    Request: ProjectionThread,
-    execute: (row) =>
-      sql`
-        INSERT INTO projection_threads (
-          thread_id,
-          project_id,
-          title,
-          model_selection_json,
-          runtime_mode,
-          interaction_mode,
-          branch,
-          worktree_path,
-          latest_turn_id,
-          created_at,
-          updated_at,
-          archived_at,
-          latest_user_message_at,
-          pending_approval_count,
-          pending_user_input_count,
-          has_actionable_proposed_plan,
-          deleted_at
-        )
-        VALUES (
-          ${row.threadId},
-          ${row.projectId},
-          ${row.title},
-          ${JSON.stringify(row.modelSelection)},
-          ${row.runtimeMode},
-          ${row.interactionMode},
-          ${row.branch},
-          ${row.worktreePath},
-          ${row.latestTurnId},
-          ${row.createdAt},
-          ${row.updatedAt},
-          ${row.archivedAt},
-          ${row.latestUserMessageAt},
-          ${row.pendingApprovalCount},
-          ${row.pendingUserInputCount},
-          ${row.hasActionableProposedPlan},
-          ${row.deletedAt}
-        )
-        ON CONFLICT (thread_id)
-        DO UPDATE SET
-          project_id = excluded.project_id,
-          title = excluded.title,
-          model_selection_json = excluded.model_selection_json,
-          runtime_mode = excluded.runtime_mode,
-          interaction_mode = excluded.interaction_mode,
-          branch = excluded.branch,
-          worktree_path = excluded.worktree_path,
-          latest_turn_id = excluded.latest_turn_id,
-          created_at = excluded.created_at,
-          updated_at = excluded.updated_at,
-          archived_at = excluded.archived_at,
-          latest_user_message_at = excluded.latest_user_message_at,
-          pending_approval_count = excluded.pending_approval_count,
-          pending_user_input_count = excluded.pending_user_input_count,
-          has_actionable_proposed_plan = excluded.has_actionable_proposed_plan,
-          deleted_at = excluded.deleted_at
-      `,
-  });
-
-  const getProjectionThreadRow = SqlSchema.findOneOption({
-    Request: GetProjectionThreadInput,
-    Result: ProjectionThreadDbRow,
-    execute: ({ threadId }) =>
-      sql`
-        SELECT
-          thread_id AS "threadId",
-          project_id AS "projectId",
-          title,
-          model_selection_json AS "modelSelection",
-          runtime_mode AS "runtimeMode",
-          interaction_mode AS "interactionMode",
-          branch,
-          worktree_path AS "worktreePath",
-          latest_turn_id AS "latestTurnId",
-          created_at AS "createdAt",
-          updated_at AS "updatedAt",
-          archived_at AS "archivedAt",
-          latest_user_message_at AS "latestUserMessageAt",
-          pending_approval_count AS "pendingApprovalCount",
-          pending_user_input_count AS "pendingUserInputCount",
-          has_actionable_proposed_plan AS "hasActionableProposedPlan",
-          deleted_at AS "deletedAt"
-        FROM projection_threads
-        WHERE thread_id = ${threadId}
-      `,
-  });
-
-  const listProjectionThreadRows = SqlSchema.findAll({
-    Request: ListProjectionThreadsByProjectInput,
-    Result: ProjectionThreadDbRow,
-    execute: ({ projectId }) =>
-      sql`
-        SELECT
-          thread_id AS "threadId",
-          project_id AS "projectId",
-          title,
-          model_selection_json AS "modelSelection",
-          runtime_mode AS "runtimeMode",
-          interaction_mode AS "interactionMode",
-          branch,
-          worktree_path AS "worktreePath",
-          latest_turn_id AS "latestTurnId",
-          created_at AS "createdAt",
-          updated_at AS "updatedAt",
-          archived_at AS "archivedAt",
-          latest_user_message_at AS "latestUserMessageAt",
-          pending_approval_count AS "pendingApprovalCount",
-          pending_user_input_count AS "pendingUserInputCount",
-          has_actionable_proposed_plan AS "hasActionableProposedPlan",
-          deleted_at AS "deletedAt"
-        FROM projection_threads
-        WHERE project_id = ${projectId}
-        ORDER BY created_at ASC, thread_id ASC
-      `,
-  });
-
-  const deleteProjectionThreadRow = SqlSchema.void({
-    Request: DeleteProjectionThreadInput,
-    execute: ({ threadId }) =>
-      sql`
-        DELETE FROM projection_threads
-        WHERE thread_id = ${threadId}
-      `,
-  });
-
-  const upsert: ProjectionThreadRepository["Service"]["upsert"] = (row) =>
-    upsertProjectionThreadRow(row).pipe(
-      Effect.mapError(toPersistenceSqlError("ProjectionThreadRepository.upsert:query")),
-    );
-
-  const getById: ProjectionThreadRepository["Service"]["getById"] = (input) =>
-    getProjectionThreadRow(input).pipe(
-      Effect.mapError(toPersistenceSqlError("ProjectionThreadRepository.getById:query")),
-    );
-
-  const listByProjectId: ProjectionThreadRepository["Service"]["listByProjectId"] = (input) =>
-    listProjectionThreadRows(input).pipe(
-      Effect.mapError(toPersistenceSqlError("ProjectionThreadRepository.listByProjectId:query")),
-    );
-
-  const deleteById: ProjectionThreadRepository["Service"]["deleteById"] = (input) =>
-    deleteProjectionThreadRow(input).pipe(
-      Effect.mapError(toPersistenceSqlError("ProjectionThreadRepository.deleteById:query")),
-    );
-
-  return {
-    upsert,
-    getById,
-    listByProjectId,
-    deleteById,
-  } satisfies ProjectionThreadRepository["Service"];
-});
-
-export const layer = Layer.effect(ProjectionThreadRepository, make);
